backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-anymail[mailgun]==14.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
Pillow==12.1.1
python-dotenv==1.1.0
PyMySQL==1.1.1
cryptography==46.0.5
sqlparse==0.5.5
google-generativeai==0.8.5
gunicorn==23.0.0
pywebpush==2.3.0
